refactor(scraping): Extract message sender
